rename package from com.google to com.google.cloud.pso, added a few edits into README.md
