+++ examples/cloud-datafusion-functions-plugins/src/main/java/com/google/cloud/pso/common/GCPReferenceSinkConfig.java
@@ -13,8 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.google.common;
+package com.google.cloud.pso.common;
 
+import com.google.cloud.pso.common.GCPConfig;
 import io.cdap.cdap.api.annotation.Description;
 import io.cdap.cdap.api.annotation.Macro;
 import io.cdap.cdap.api.annotation.Name;
