+++ examples/cloud-datafusion-functions-plugins/src/main/java/com/google/cloud/pso/common/GCPUtils.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.google.common;
+package com.google.cloud.pso.common;
 
 import com.google.auth.oauth2.ServiceAccountCredentials;
 import java.io.File;
