+++ examples/cloud-datafusion-functions-plugins/src/main/java/com/google/cloud/pso/plugin/CopyTableAction.java
@@ -13,9 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.google.plugin;
+package com.google.cloud.pso.plugin;
 
-import com.google.functions.InsertOverwriteFunction;
+import com.google.cloud.pso.functions.InsertOverwriteFunction;
 import io.cdap.cdap.api.TxRunnable;
 import io.cdap.cdap.api.annotation.Description;
 import io.cdap.cdap.api.annotation.Macro;
